Make unit tests compatible with Python 3.13
Python 3.13 has changed the behaviour of the `xml.dom.minidom` module:
double quotes used to be escaped everywhere, but they are now only
escaped inside attributes.

This was done by the following commit.
https://github.com/python/cpython/commit/154477be722ae5c4e18d22d0860e284006b09c4f

As a result, the unit tests fail when executed on Python 3.13.

Fix this issue in the following way:
1. In all 88 reference-output XML files, replace all occurrences of
   `&quot;` with `"`.
2. Change `tests/conftest.py` to force-replace all occurrences of
   `&quot;` with `"` in the test output before comparing it to the
   reference output. This ensures that the subsequent equality assertion
   still holds on older versions of Python.

This solution works because the test data does not contain any instances
of `&quot;` inside XML attributes.

Change-Id: Ied875014b1bb5d8225943303858658648efc12ad